+++ src/screen/BulletinBoard.jsx
@@ -1,7 +1,6 @@
 import Navibar from "@/components/Navibar.jsx";
 import SideBar from "@/components/SideBar.jsx";
 import Inquiry from "@/components/Inquiry.jsx";
-import Board from "@/components/Board.jsx";
 
 function BulletinBoard() {
     return (
